feat: 타입 정의 이동 및 세션 관리 추가
